pcbgenerator/requirements-dev.txt
freetype-py==2.5.1
svg.path==7.0
fonttools==4.59.0
pytest==8.4.1
pytest-timeout==2.4.0
cairosvg==2.8.2
pillow==11.3.0
shapely==2.1.1
